Added first part of full image detection (V1)
Updated model.py to enable memory growth so that model does not take all VRAM available.
Updated datasetDivider.py to be able to use non-square divisions and also to compute max pixel length of an axis limited by number of div.
Updated nephrology.py to divide cleaned image of cortex detection in order to perform main detection on it.
Updated Inference Notebook to support customization of parameters of the cleaned image's divisions.